+++ src/interface/src/swig/fortrilinosFORTRAN_wrap.cxx
@@ -1326,14 +1326,14 @@ SWIGEXPORT void _wrap_TrilinosEigenSolver_setup_solver(SwigClassWrapper const *f
 }
 
 
-SWIGEXPORT size_t _wrap_TrilinosEigenSolver_solve(SwigClassWrapper const *farg1, SwigArrayWrapper *farg2, SwigClassWrapper const *farg3, SwigArrayWrapper *farg4) {
-  size_t fresult ;
+SWIGEXPORT int _wrap_TrilinosEigenSolver_solve(SwigClassWrapper const *farg1, SwigArrayWrapper *farg2, SwigClassWrapper const *farg3, SwigArrayWrapper *farg4) {
+  int fresult ;
   ForTrilinos::TrilinosEigenSolver *arg1 = (ForTrilinos::TrilinosEigenSolver *) 0 ;
   std::pair< ForTrilinos::TrilinosEigenSolver::SC *,std::size_t > arg2 ;
   Teuchos::RCP< ForTrilinos::TrilinosEigenSolver::MultiVector > *arg3 = 0 ;
   std::pair< int *,std::size_t > arg4 ;
   Teuchos::RCP< ForTrilinos::TrilinosEigenSolver::MultiVector > tempnull3 ;
-  size_t result;
+  int result;
   
   SWIG_check_nonnull(*farg1, "ForTrilinos::TrilinosEigenSolver const *", "TrilinosEigenSolver", "ForTrilinos::TrilinosEigenSolver::solve(std::pair< ForTrilinos::TrilinosEigenSolver::SC *,std::size_t >,Teuchos::RCP< ForTrilinos::TrilinosEigenSolver::MultiVector > &,std::pair< int *,std::size_t >) const", return 0);
   arg1 = static_cast< ForTrilinos::TrilinosEigenSolver * >(farg1->cptr);
@@ -1348,7 +1348,7 @@ SWIGEXPORT size_t _wrap_TrilinosEigenSolver_solve(SwigClassWrapper const *farg1,
     try
     {
       // Attempt the wrapped function call
-      result = (size_t)((ForTrilinos::TrilinosEigenSolver const *)arg1)->solve(arg2,*arg3,arg4);
+      result = (int)((ForTrilinos::TrilinosEigenSolver const *)arg1)->solve(arg2,*arg3,arg4);
     }
     catch (const std::range_error& e)
     {
@@ -1365,7 +1365,7 @@ SWIGEXPORT size_t _wrap_TrilinosEigenSolver_solve(SwigClassWrapper const *farg1,
       SWIG_exception_impl("ForTrilinos::TrilinosEigenSolver::solve(std::pair< ForTrilinos::TrilinosEigenSolver::SC *,std::size_t >,Teuchos::RCP< ForTrilinos::TrilinosEigenSolver::MultiVector > &,std::pair< int *,std::size_t >) const", SWIG_UnknownError, "An unknown exception occurred", return 0);
     }
   }
-  fresult = static_cast< size_t >(result);
+  fresult = static_cast< int >(result);
   return fresult;
 }
